refactor(profile, nft): add mock service and extract NFT row components
- Add MockProfileService actor for previews and tests
  - Implement ProfileServiceProtocol
  - Implement NftServiceProtocol
  - Provide sample profile and NFT test data

- Extract MyNFTsRowView into a separate file
  - Move NFTImageView
  - Move NFTImagePlaceholder
  - Move RatingStarsView
  - Move ProfileNft helper extension

- Remove duplicated NFT row UI from MyNFTsView
- Update MyNFTsView preview to use MockProfileService
- Register new source files in Xcode project
